Revert "Login screen (#8)"
This reverts commit 148bce127d42e8c9fc7729cc86a75bb2ccdb44f8.
