+++ .gitignore
@@ -141,7 +141,7 @@ playground.xcworkspace
 *.xcuserdata
 
 ## Cocoapods ##
-Pods/
+ownCloud/Pods/
 
 # End of https://www.gitignore.io/api/xcode,swift,macos,objective-c
