--- ownCloud/Resources/Localizable.strings
@@ -1,25 +1,0 @@
-/* 
-  Localizable.strings
-  ownCloud
-
-  Created by Pablo Carrascal on 13/03/2018.
-  Copyright © 2018 ownCloud GmbH. All rights reserved.
-*/
-"Edit Server" = "Edit Server";
-"Add Server" = "Add Server";
-"Server URL" = "Server URL";
-"https://example.com" = "https://example.com";
-"Continue" = "Continue";
-"Name" = "Name";
-"Example Server" = "Example Server";
-"Show Certificate Details" = "Show Certificate Details";
-"Connect" = "Connect";
-"Delete Authentication Data" = "Delete Authentication Data";
-"Authentication" = "Authentication";
-"Username" = "Username";
-"Password" = "Password";
-"Certificate Details" = "Certificate Details";
-"Cancel" = "Cancel";
-"Approve" = "Approve";
-"Error" = "Error";
-"Review Connection" = "Review Connection";
