+++ ownCloud/ownCloud/AppDelegate.swift
