+++ ownCloud/ownCloud/Bookmarks/BookmarkManager.swift
