+++ ownCloud/ownCloud/Resources/Assets.xcassets/AppIcon.appiconset/Contents.json
