+++ ownCloud/ownCloud/Resources/Assets.xcassets/AppIcon.appiconset/ownCloud_ios_beta_icon-20.png
