+++ ownCloud/ownCloud/Resources/Assets.xcassets/Contents.json
