+++ ownCloud/ownCloud/Resources/Assets.xcassets/owncloud-primary-small.imageset/Contents.json
