+++ ownCloud/ownCloud/Resources/Assets.xcassets/owncloud-primary-small.imageset/owncloud-primary-small.pdf
