+++ ownCloud/ownCloud/Server List/ServerListBookmarkCell.swift
