+++ ownCloud/ownCloud/Server List/ServerListTableViewController.swift
@@ -134,9 +134,10 @@ class ServerListTableViewController: UITableViewController {
 
 	// MARK: - Actions
 	@IBAction func addBookmark() {
+		let bookmark = OCBookmark(for: URL(string: "https://demo.owncloud.org"))
 
-        let viewController : BookmarkViewController = BookmarkViewController(style: UITableViewStyle.grouped)
-        self.navigationController?.pushViewController(viewController, animated: true)
+		BookmarkManager.sharedBookmarkManager.addBookmark(bookmark!)
+
 		updateNoServerMessageVisibility()
 	}
