+++ ownCloud/ownCloud/Tools/Log.swift
