+++ ownCloud/ownCloud/UI Elements/ActionButton.swift
